Jetzt funktioniert der Login

+++ Umsetzung/Adlatus_Laravel/database/migrations/2017_11_17_132107_create_table_therapeuts.php
@@ -14,7 +14,7 @@ class CreateTableTherapeuts extends Migration
     public function up()
     {
         Schema::create('therapeuts', function(Blueprint $table) {
-            $table->integer('sozNr')->primary();
+            $table->bigInteger('sozNr')->primary();
             $table->string('vorname');
             $table->string('nachname');
             $table->string('email')->nullable();
